chore: restore x-ipfs-path
